requirements.txt
arabic-reshaper==2.1.3
asgiref==3.2.10
boto3==1.9.96
botocore==1.12.96
certifi==2020.6.20
chardet==3.0.4
dj-database-url==0.5.0
Django==3.1.2
django-crispy-forms==1.12.0
django-grappelli==2.15.1
django-heroku==0.3.1
django-mathfilters==1.0.0
django-multiselectfield==0.1.12
django-phone-field==1.8.1
django-restframework==0.0.1
django-storages==1.7.1
django-taggit==1.4.0
django-taggit-autosuggest==0.3.8
django-widget-tweaks==1.4.8
djangorestframework==3.12.4
docutils==0.14
future==0.18.2
gunicorn==19.9.0
html5lib==1.1
idna==2.10
jmespath==0.9.3
octicons==2.1.0
Pillow==8.2.0
psycopg2==2.9.1
psycopg2-binary==2.9.1
PyMySQL==1.0.2
PyPDF2==1.26.0
python-bidi==0.4.2
python-dateutil==2.8.0
pytz==2020.1
razorpay==1.2.0
reportlab==3.5.67
requests==2.24.0
s3transfer==0.2.0
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.10
validate-email==1.3
webencodings==0.5.1
whitenoise==4.1.2
xhtml2pdf==0.2.5
zipp==3.3.0
django-taggit-autosuggest==0.3.8
django-grappelli==2.15.1
